Add Metric/SingleMetric/MultiMetric
These classes act as the ferry between raw data and prometheus exporter
text provided by this exporter. This change introduces two types of
metrics behind an abstract class.

The Metric abstract class is the base object that can be required via
polymorphism. This is due to the way prometheus allows for either
single-valued metrics, or multi-valued metrics.

SingleMetric is the basic type of metric which is solely a key-value
pair shows like:
`metric_name value`

MultiMetric is the compound type of metric which allows for labeling of
the same metric with several breakdowns like:
`metric_name{foo=1,bar=a} value`

As both are valid in all types of exporter, the base Metric class will
be used throughout to specify when a metric is needed.