Create core server logic for exporter
Adds ktor server logic like responding to / and /metrics. Also adds debug logic which allows us to serve qbt traffic through a file. This change sets the groundwork for the exporter server with a fully working e2e pipeline for serving metrics to prometheus. Initial implementation is not meant to be production ready, but rather a platform prototype and iterate on.