Refresh in-memory settings when settings file changes
This allows users to modify the settings file and not have to restart the exporter in order for those settings to take effect. Note, this implementation is dependent on JVM due to the way the OS reports the time. At the slowest, the JVM will return time precision in seconds, requiring users to make sequantical settings file changes over 1 second long which is acceptable for 99% of users.