Remove thread block when server is running
Instead of blocking the thread when running the ktor server, we return immediately to allow the thread itself to manage its own lifecycle. This enables unit tests to manage the server's lifecycle more consistently instead of relying on system threads to kill the process.

This commit fixes the ExporterServerImplTest flakiness.