Fail ExporterServerImplTest early if server cannot start
Previously, if the ExporterServerImplTest tests ran when an existing server was running, the test behavior was undefined as it was possible the tests could utilize the existing server instance to test behavior. Now, the test will fail fast if server startup contains an exception (where previously it would silently continue).