Abstract maindata collector fetching logic to CollectorSettingsUtils
This change makes CollectorSettingsUtils into an interface which allows
us to mock its behavior in tests. Specifically, the logic to fetch
enabled maindata collectors from the settings file.

This change also uses a lazyRelay to update the settings field so that
we're always feteching the most recent settings from disk if the file
has changed by the user.