Add LazyRelayer to SettingsManager
Instead of relying on downstream users of settings to implement a
LazyRelayer, the settings manager will only provide settings via a
relayer. This abstracts relaying logic up to the settings manager so
downstream classes won't need to re-implement the logic. This change
also deprecates the use of #get as it will no longer be needed (in favor
of #relay).