Add MaindataCollectorContainer and associated settings.
The container interface will apply to the maindata collector enums that house the the collectors. The goal is to have the export server call the enum aggregate method with its settings instead of parsing the settings beforehand and calling individual collector methods. In this model, we move the responsibility of producing metrics to the collector itself instead of relying on an external class knowing when it should enable or disable a collector.